OptionsParser: parse the cookiesfrombrowser string
When used as a library, the argument string is stored as-is, which
later results in it getting exploded into a list of characters, rather
than the tuple of values it is supposed to be.

This adds a callback to parse the cookiesfrombrowser string format and
store it as a tuple in the parser values.

Fixes #10196.